@GitHK review: move helper
